Refactor pluck
