+++ lib/__tests__/pluck.test.js
@@ -1,36 +1,23 @@
-import { prop, pluck } from '..'
+import { pluck } from '..'
 
 describe('pluck', () => {
-  const obj = {
-    fruits: ['apple', 'orange', 'banana'],
-    total: 138,
-  }
+  const data = [
+    {
+      city: 'New York',
+      state: 'NY',
+    },
+    {
+      city: 'San Francisco',
+      state: 'CA',
+    },
+    {
+      city: 'Portland',
+      state: 'OR',
+    },
+  ]
 
   it('returns pluck', () => {
     expect.assertions(1)
-    expect(pluck('fruits', obj)).toStrictEqual(obj.fruits)
-  })
-})
-
-describe('prop', () => {
-  const data = {
-    name: 'Bob',
-    age: 40,
-  }
-
-  const name = prop('name')
-
-  it('returns prop value', () => {
-    expect.assertions(1)
-    expect(name(data)).toStrictEqual('Bob')
-  })
-
-  let rename = name(data)
-  rename = 'Tom'
-
-  it('no side-effects', () => {
-    expect.assertions(2)
-    expect(data.name).toStrictEqual('Bob')
-    expect(rename).toStrictEqual('Tom')
+    expect(pluck('state', data)).toStrictEqual(['NY', 'CA', 'OR'])
   })
 })
