+++ lib/pluck.js
@@ -1,5 +1,7 @@
+import curry from './curry'
+import map from './map'
 import prop from './prop'
 
-const pluck = prop
+const pluck = curry((key, list) => map(item => prop(key, item), list))
 
 export default pluck
